More spaces in solver, build.groovy zap 00
The build.groovy had the name for the output file for test 1 to
be output_001. Unfortunately, test 10 was output_0010 (and likely
was intended to be output_010).

I zapped the leading 00. Now the output files are sync'ed with the
test names:
wrf_testcase_1.sh and output_1
wrf_testcase_10.sh and output_10